axf/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==1.11.7
django-debug-toolbar==2.0
django-redis==4.10.0
idna==2.8
Pillow==6.2.0
pkg-resources==0.0.0
pycryptodomex==3.7.2
PyMySQL==0.9.3
pytz==2019.2
redis==3.3.10
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
uWSGI==2.0.18
